Refactor package structure and update Ui methods
Files like `Parser.java` and `Storage.java` are scattered across
different packages, leading to inconsistencies in project structure.

To maintain a clean and organized codebase, files related to command
execution and components should be grouped in one package.

The files `Parser.java`, `Storage.java`, and `UnmarkCommand.java` have
been moved to the `components` package. Additionally, the methods in
`Ui.java` that previously printed messages directly have been refactored
to return strings instead. This enhances flexibility and supports better
unit testing by decoupling UI logic from direct console printing.

This approach improves maintainability and allows for better
modularization of the system.
